Extends cmd-line, some little fixes
This commit extends command line by
 + -s/--simulate - do not write files or create directories
 + -q/--quiet    - do not print data summary (concerns Logger::DEBUG only)
 + -l/--level    - set logger level

Fixes
 + removes non-joke
 + corrects spelling mistake
 + tell tabula to also export line breaks
 + extr_helper
   - copy_sry_file `exit` -> `return`
   - adds missing `echo`

Furthermore updates `.gitignore`.

+++ studienplan5.rb
@@ -129,11 +129,11 @@ OptionParser.new do |opts|
         $options[:classes] = j
     end
 
-    opts.on("-o", "--output NAME", "Specify output target, if ends with slash, will be output directory. If not, will be name of calendar dir and suffix for JSON files.") do |o|
+    opts.on("-o", "--output NAME", "Specify output target, if ends with slash, will be output directory. If not, will be name of calendar dir and prefix for JSON files.") do |o|
         $options[:output] = o
     end
 
-    opts.on("-k", "--disable-json-object-keys", "Stringify hash keys.") do |jok|
+    opts.on("-k", "--disable-json-object-keys", "Stringify keys that are not strings.") do |jok|
         $options[:no_jok] = jok
     end
 
@@ -149,12 +149,24 @@ OptionParser.new do |opts|
         $options[:cal_dir] = cal_dir
     end
 
-    opts.on("-u", "--disable-unified", "Do not create files that contain all parent events recursively.") do |u| # Caution, you'll be the next company we buy!
+    opts.on("-u", "--disable-unified", "Do not create files that contain all parent events recursively.") do |u|
         $options[:no_unified] = u
     end
 
     opts.on("-a", "--disable-apache-config", "Do not export .htaccess and other Apache-specific customizations.") do |no_apache|
         $options[:no_apache] = no_apache
+    end
+
+    opts.on("-s", "--simulate", "Simulate, do not write files or create directories.") do |simulate|
+        $options[:simulate] = simulate
+    end
+
+    opts.on("-q", "--quiet", "Do not print data.") do |quiet|
+        $options[:quiet] = quiet
+    end
+
+    opts.on("-l", "--level [LEVEL]", {fatal: Logger::FATAL, error: Logger::ERROR, warn: Logger::WARN, info: Logger::INFO, debug: Logger::DEBUG}, "Log level") do |level|
+        $logger.level = level
     end
 
     opts.on("-h", "--help", "Print this help.") do |h|
@@ -175,7 +187,7 @@ if outp
         data_file = outp + data_file
         classes_file = outp + classes_file
 
-        Dir.mkdir(outp) unless Dir.exists?(outp)
+        Dir.mkdir(outp) unless Dir.exists?(outp) or $options[:simulate]
     else
         ical_dir = outp
         data_file = outp + ".data.json"
@@ -357,7 +369,7 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
 
                 rowPart.map.with_index do |element, k|
 
-                    $logger.debug "row #{i}, part #{j}, element #{k}"
+                    $logger.debug "row #{i}, part #{j}, element #{k}" unless $options[:no_progress]
 
                     # As mentioned above step 1
                     cw = ( cw = plan[0][j][k] ) ? cw.text : cw
@@ -750,17 +762,19 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
             end # parts iteration
         end # skip first two rows
     end # rows iteration
-    $logger.debug "Data following..."
-    mapped.each do |row|
-        if row
-            row.each do |part|
-                if part
-                    part.each do |element|
-                        if element and element.length > 0
-                            element.each do |subelement|
-                                $logger.debug subelement.to_s
+    unless $options[:quiet]
+        $logger.debug "Data following..."
+        mapped.each do |row|
+            if row
+                row.each do |part|
+                    if part
+                        part.each do |element|
+                            if element and element.length > 0
+                                element.each do |subelement|
+                                    $logger.debug subelement.to_s
+                                end
+                                $logger.debug "--"
                             end
-                            $logger.debug "--"
                         end
                     end
                 end
@@ -778,8 +792,12 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
 
         $logger.debug "Writing JSON data file \"%s\"" % data_file
 
-        File.open(data_file, "w+") do |datafile|
-            datafile.puts $options[:json_pretty] ? JSON.pretty_generate(json_data) : JSON.generate(json_data)
+        if $options[:simulate]
+            $logger.info "Would write #{data_file}"
+        else
+            File.open(data_file, "w+") do |datafile|
+                datafile.puts $options[:json_pretty] ? JSON.pretty_generate(json_data) : JSON.generate(json_data)
+            end
         end
 
         $logger.info "Wrote JSON data file \"%s\"" % data_file
@@ -793,7 +811,7 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
         cal_stub.prodid = "-Christoph criztovyl Schulz//studienplan5 using icalendar-ruby//DE"
         cal_stub.add_timezone tz.ical_timezone(Time.now)
 
-        Dir.mkdir(ical_dir) unless Dir.exists?(ical_dir)
+        Dir.mkdir(ical_dir) unless Dir.exists?(ical_dir) or $options[:simulate]
 
         $logger.info "Writing unified calendars." unless no_unified
 
@@ -809,8 +827,12 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
 
             $logger.debug "Writing calendar file \"%s\"" % clazz_file
 
-            File.open(clazz_file, "w+") do |f|
-                f.puts cal.to_ical
+            if $options[:simulate]
+                $logger.info "Would write #{clazz_file}"
+            else
+                File.open(clazz_file, "w+") do |f|
+                    f.puts cal.to_ical
+                end
             end
         end
 
@@ -840,8 +862,12 @@ elsif $options[:json] or $options[:ical] or $options[:classes]
 
         $logger.debug "Writing JSON classes file \"%s\"" % classes_file
 
-        File.open(classes_file, "w+") do |datafile|
-            datafile.puts $options[:json_pretty] ? JSON.pretty_generate(json_data) : JSON.generate(json_data)
+        if $options[:simulate]
+            $logger.info "Would write #{classes_file}"
+        else
+            File.open(classes_file, "w+") do |datafile|
+                datafile.puts $options[:json_pretty] ? JSON.pretty_generate(json_data) : JSON.generate(json_data)
+            end
         end
 
         $logger.info "Wrote JSON classes file \"%s\"" % classes_file
@@ -858,16 +884,20 @@ if $options[:web] and $options[:output] and $options[:output].end_with?(?/)
     sep = File::SEPARATOR
     o = $options[:output] + sep
 
-    FileUtils.cp_r "web/.", $options[:output]
-    if not $options[:no_apache]
-        if Dir.exists? ical_dir
-            FileUtils.mv o + "indexes_header.html", ical_dir
-            FileUtils.cp o + "cover.css", ical_dir + sep + "indexes_css.css"
-        else
-            $logger.info "Target dir for icals does not exist, please specify it's name by --calendar-dir to enable custom Apache indexes style."
-        end
+    if $options[:simulate]
+        $logger.info "Would copy ./web/ to #{o}"
     else
-        FileUtils.rm [o + ".htaccess", o + "indexes_header.html"]
+        FileUtils.cp_r "web/.", o
+        if not $options[:no_apache]
+            if Dir.exists? ical_dir
+                FileUtils.mv o + "indexes_header.html", ical_dir
+                FileUtils.cp o + "cover.css", ical_dir + sep + "indexes_css.css"
+            else
+                $logger.info "Target dir for icals does not exist, please specify it's name by --calendar-dir to enable custom Apache indexes style."
+            end
+        else
+            FileUtils.rm [o + ".htaccess", o + "indexes_header.html"]
+        end
     end
 
     $logger.warn "You haven't exported classes (-d/--classes) yet but they are required by -w/--web!" unless File.exists?(o+"classes.json")
